requirements.txt
aiohttp
async-timeout
attrs
ButtonPaginator
certifi
chardet
charset-normalizer
discord-components
discord-py-interactions
discord-py-slash-command==3.0.3
discord-select-menus
discord.py
discord.py-message-components
dislash.py
gunicorn
idna
mojang
multidict
numpy
opencv-contrib-python
opencv-python
pandas
PyMySQL
python-dateutil
pytz
requests
six
typing-extensions
urllib3
yarl
